move logic and delete unused deposits file
