+++ packages/indexer/src/services/index.ts
@@ -1,4 +1,3 @@
-export * as deposits from "./deposits";
 export * as bundles from "./bundles";
 export * as spokePoolIndexer from "./spokePoolIndexer";
 export * as hubPoolIndexer from "./hubPoolIndexer";
